+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added `new_id` function
 - Deprecated `next_id` function in favor of `new_id`
 - Added `identity` and `has_identity` functions
+- Added `singleton` and `events` functions
 
 ## 1.1.0 (September 12, 2017)